split the Chatbot code into 3 files. Added Utilities folder where I put the Chatbot functions, and the Chatbot DB requests
